Dodanie wykresów

+++ divvyup/DivvyUp.Shared/Dto/ChartDto.cs
@@ -3,6 +3,6 @@
     public class ChartDto
     {
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
     }
 }
